Removed public.html, added what will you learn partial, rendered it in attend.
