+++ app/views/pages/attend.html.erb
@@ -1,2 +1,1 @@
-<h1>Pages#attend</h1>
-<p>Find me in app/views/pages/attend.html.erb</p>
+<%= render partial: "shared/what_will_you_learn" %>
