+++ app/views/shared/_what_will_you_learn.html.erb
@@ -1,0 +1,28 @@
+<div class="container marketing">
+  <div class="row">
+    <div class="container">
+      <div class="row">
+        <h1 style="text-align:center;">What will you learn</h1>
+      </div>
+      <div class="row">
+        <h2 style="text-align:center;" class="text-muted">Professional developers technologies</h2>
+        <div class="col-lg-4">
+          <img class="img-circle" src="data:image/gif;base64,R0lGODlhAQABAIAAAHd3dwAAACH5BAAAAAAALAAAAAABAAEAAAICRAEAOw==" alt="HTML" width="140" height="140">
+          <h2>HTML</h2>
+          <p>HTML is a computer language devised to allow website creation. These websites can then be viewed by anyone else connected to the Internet. It is relatively easy to learn, with the basics being accessible to most people in one sitting; and quite
+              powerful in what it allows you to create.</p>
+        </div>
+        <div class="col-lg-4">
+          <img class="img-circle" src="data:image/gif;base64,R0lGODlhAQABAIAAAHd3dwAAACH5BAAAAAAALAAAAAABAAEAAAICRAEAOw==" alt="CSS" width="140" height="140">
+          <h2>CSS</h2>
+          <p>CSS is a style language that defines layout of HTML documents. For example, CSS covers fonts, colours, margins, lines, height, width, background images, advanced positions and many other things.</p>
+        </div>
+        <div class="col-lg-4">
+          <img class="img-circle" src="data:image/gif;base64,R0lGODlhAQABAIAAAHd3dwAAACH5BAAAAAAALAAAAAABAAEAAAICRAEAOw==" alt="JavaScript" width="140" height="140">
+          <h2>JavaScript</h2>
+          <p>JavaScript is a programming language used to make web pages interactive. JavaScript is responsible for the "behavior" of your Website. It's how HTML elements and CSS style animate and move around on the page.</p>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
