+++ spec/views/pages/shared/_what_will_you_learn.html.erb_spec.rb
@@ -1,0 +1,13 @@
+require 'rails_helper'
+
+describe "pages/public.html.erb" do
+  it "contains \"Professional developers technologies\"" do
+    render
+    response.should have_text("Professional developers technologies")
+  end
+
+  it "contains \"JavaScript is a programming language\"" do
+    render
+    response.should have_text("JavaScript is a programming language")
+  end
+end
